pyproject.toml
[tool.poetry]
name = "tasks"
version = "0.1.0"
description = ""
authors = ["shencar <barak.shencar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.3"
uvicorn = "^0.29.0"
google-api-python-client = "^2.24.0"
google-auth-oauthlib = "^1.2.0"
flask = "^3.0.3"
aiohttp = "^3.9.5"
google-api-core = "^2.19"
pytest = "^8.2.0"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pytest-asyncio = "^0.23.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
